Refactor djaypro: extract DB helpers into mediadb module (#1696)
